requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
